pyproject.toml
[project]
name = "NervaAPI"
version = "1.0.0"
description = "NervaAPI is a REST API interface for the Nerva (XNV) blockchain."
license = "GPL-3.0"
authors = [
  { name="Sayan Bhattacharyya", email="sayan@sn1f3rt.dev" },
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp==3.10.10",
    "motor==3.6.0",
    "nerva-py==1.1.1",
    "quart-rate-limiter==0.10.0",
    "quart==0.20.0",
    "schedule==1.2.2",
    "validators==0.34.0",
]

[project.optional-dependencies]
speed = [
    "uvloop==0.22.1",
]

[dependency-groups]
dev = [
    "pre-commit==4.5.1",
    "ruff==0.14.14",
]

[tool.ruff]
line-length = 85
target-version = "py313"

[tool.ruff.lint]
select = ["I"]

[tool.ruff.lint.isort]
section-order = ["future", "typing", "standard-library", "third-party", "first-party", "local-folder"]
force-wrap-aliases = true
combine-as-imports = true
length-sort = true

[tool.ruff.lint.isort.sections]
"typing" = ["typing"]
